Made time int64 and cleaned some truncations.

+++ Microsoft.Toolkit.Uwp.Input.Gaze/GazeTargetItem.cpp
@@ -58,7 +58,8 @@ void GazeTargetItem::RaiseProgressEvent(GazeProgressState state)
             auto control = safe_cast<Control^>(TargetElement);
 
             auto transform = control->TransformToVisual(_feedbackPopup);
-            auto bounds = transform->TransformBounds(*ref new Rect(*ref new Point(0, 0), *ref new Size(control->ActualWidth, control->ActualHeight)));
+            auto bounds = transform->TransformBounds(*ref new Rect(*ref new Point(0, 0), 
+                *ref new Size(safe_cast<float>(control->ActualWidth), safe_cast<float>(control->ActualHeight))));
             auto rectangle = safe_cast<Rectangle^>(_feedbackPopup->Child);
 
             if (state == GazeProgressState::Progressing)
